DeviceMgr: First implementation based on CfgMgr32 API
This is the first commit that provides a .NET native access to device information on a Windows System, tested from Windows XP until Windows 11. It uses the CfgMgr32.DLL API to query the devices in the system and get properties of the devices. This is a direct way to obtain data, instead of using WMI.

To test the software and dump the device list to the screen, run the `DeviceInfoDump` tool.

Issue: DOTNET-700