DeviceInstance: Provide refresh mechanism for a device instance
The instances are cached in memory, sortable by their device identifier, which is documented by Windows to be unique for an index (but not to interpret the identifier, thus this makes a good key). This allows subsequent calls to `GetRoot()` to return the same value, assuming that it hasn't changed.

Because of the new assumption, we add a new method called `Refresh()` that can be used on an existing tree, or part of a tree, to refresh the device nodes. We don't need to call `GetRoot()` a second time to iterate the entire tree, if we're only interested in a part of it. It will get the status again, reset the properties (so reading the properties again might return new values), and repopulate the children, adding/removing nodes as necessary.

Issue: DOTNET-707