DeviceInstance: Get a list of device drivers
Getting the tree, which uses `CM_Get_Child()` and `CM_Get_Sibling()` only show the devices that are currently present. To get a complete list, one must get the list, and then locate each node explicitly.

Warnings were removed from logging that are normal for when a device isn't present.

As we now get the list, it's difficult to manage disposing the large number of objects. We actually don't do anything in dispose, except to dispose the handles, which don't clean anything up. So the `Dispose()` functions were also completely removed. This works much better with caching also.

Issue: DOTNET-708