DeviceInstance: Get parents when reading the list
The list now reads the parents, and this is used to build the tree. You can now get the list with `GetList()`, which also returns devices not attached, and then call `GetRoot()` to get the root node, which won't reiterate, but just returns the root node.

The `PopulateChildren()` was fixed to make sure that on overwrite, the children are removed if there are none found, instead of leaving data behind as a result of adding parents.

Issue: DOTNET-709