DeviceMgr: Optimize string based CfgMgr API
We already follow guidelines given by [1] and don't use `StringBuilder`. However, for small allocations, it's about 10x faster if we can allocate on the stack (also suggested by [1]). Write an implementation specific for the framework we're targetting for best performance. An `ArrayPool` is used for .NET Core [2], and just allocating on the heap otherwise.

[1] https://learn.microsoft.com/en-us/dotnet/fundamentals/code-analysis/quality-rules/ca1838
[2] https://learn.microsoft.com/en-us/dotnet/standard/native-interop/best-practices

The classes `DeviceInstance` and `DeviceProperty` is simplified.

Issue: DOTNET-712