DeviceInstance: Key the cache based on the handle, not the device identifier
The API appears set up in such a way that the CfgMgr DevInst handles are global, because they can never be freed. Exploit this fact and cache the handles instead of the names to make caching more efficient (because the entire CfgMgr32 API appears to be based on handles, not strings), i.e. we convert strings to handles, but traversing siblings, children and parents are done via the handles.

This also makes rebuilding the tree via `GetList()` faster, as we don't need to query for the strings all the time.

If we do have a name, such as from `GetList()`, then we don't need to query this again. Thus optimise the code and just use the name we have without querying.

Finally, because we're reworking the names, the name of the object may now also be null, as in previous PRs, and the output debug strings is now corrected.

Issue: DOTNET-723