Legacy Test: Ensure directories are properly under `legacy`
If the option `-c <Config>` is not provided on the command line, there is no configuration and files are put in the wrong directory. Follow the .NET SDK properties files for .NET SDK 7.0.302 and set the default configuration to be `Debug` if it is not provided.

Update the the path of the NuGet temporary files from `legacy` to under `obj` to be similar to the defaults in the .NET SDK.

Issue: DOTNET-827