DeviceManager: Reset internal cache on getting the device list
When calling `GetList()` should reset the cache. In this version, resetting the cache will cause reloading all devices instead of relying on the data that already exists ensuring changes to the tree are updated and reflected properly.

Code was refactored so that on `Refresh()` we don't clear the list and then add back, we set a flag, query, and then remove those elements that weren't found after the refresh. If we ever want to keep the phantom objects by doing a refresh, clearing the cache above actually fixes a problem (getting the list with all objects, then getting the list with no objects and cached, then refreshing returns incorrect results).

Issue: DOTNET-1032