DeviceManager: Refresh now properly detects inserted devices
When devices were removed, they remained in the cache, so that when they were redetected, they weren't properly updated. Because the refresh after a device is removed resulted in the children removed, and then a refresh when devices are inserted didn't update, the refresh wouldn't detect new devices.

Issue: DOTNET-1033