DeviceManager: Remove internal usage of SafeDevInst
The `IntPtr` that is provided is the handle for the `CM_*` interfaces for the local node. These can't be freed, so there's really no reason to have them as a `SafeHandle`. This reduces the load on the GC and should make it slightly faster.

The only objects that are created during `Refresh` now are the enumerator when parsing the device tree.

Issue: DOTNET-1036