Investigate missing devices
Provide a C++ test program (compiled on Windows 11, with VS2022) that does the same thing as the .NET program to confirm behaviour.

On one machine of mine with Windows 11, the function `CM_Get_Device_ID_List` doesn't return all devices. If we enumerate with `CM_Get_Child` and `CM_Get_Sibling`, then extra devices might appear (but the phantom devices disappear).

Document this as a potential problem in the README.md, but don't provide a fix. A fix is not provided as it may introduce other problems, especially if devices are removed between calls (i.e. this is why I don't try to cache data).

Issue: DOTNET-1043